fix : 저장시 오류가 생기던 부분 수정

+++ work/services/before_install_services.py
@@ -1,5 +1,6 @@
 import string
 import random
+import datetime
 
 from django.utils import timezone
 from django.http import JsonResponse
@@ -41,33 +42,11 @@ def assign_cm(request):
         f"/work/measure_before_install/{doc.urlCode}/",
     )
     cm_phone = cm.phone
-    sms_send(link, [cm_phone], 2)
-    # checklist_ids = request.POST.getlist("checklist_id")
-    # message_list = []
-    # for checklist_id in checklist_ids:
-    #     target = BeforeInspectionResult.objects.get(
-    #         before_install_checklist_id=doc,
-    #         before_inspection_item_id=BeforeInspectionItem.objects.get(
-    #             pk=checklist_id,
-    #         ),
-    #     )
-    #     measure = target.before_measure.last()
-    #     if measure.isCM:
-    #         message_list.append(
-    #             {
-    #                 "content": measure.content,
-    #                 "img": list(
-    #                     map(
-    #                         lambda x: x.img,
-    #                         list(
-    #                             measure.before_measure_imgs.all(),
-    #                         ),
-    #                     )
-    #                 ),
-    #             }
-    #         )
-    # image_send(message_list, cm_phone)
-    # 문자 전송 페이지 만들기
+    due_date = datetime.datetime.strptime(
+        request.POST.get("expired_date"), "%Y-%m-%dT%H:%M"
+    )
+
+    sms_send(link, [cm_phone], 2, due_date.strftime("%Y년 %m월 %d일 %I:%M %p"))
 
 
 def get_require_users():
@@ -255,25 +234,28 @@ def review_before_install_checklist_service(request, type, pk):
         inspectionResults = checklist.before_inspection_result.filter(result="2")
         for inspectionResult in inspectionResults:
             lastMeasure = inspectionResult.before_measure.last()
-            if lastMeasure.isCM:
-                newMeasure = inspectionResult.before_measure.create(
-                    content=request.POST[f"{inspectionResult.pk}-content"],
-                )
-                images = request.FILES.getlist(f"{inspectionResult.pk}-images[]")
-                for img in images:
-                    newMeasure.before_measure_imgs.create(img=img)
-            else:
-                lastMeasure.content = request.POST[f"{inspectionResult.pk}-content"]
-                lastMeasure.save()
-                before_images = list(
-                    request.POST.getlist(f"{inspectionResult.pk}-images-preloaded[]")
-                )
-                for measure_img in lastMeasure.before_measure_imgs.all():
-                    if str(measure_img.pk) not in before_images:
-                        measure_img.delete()
-                images = request.FILES.getlist(f"{inspectionResult.pk}-images[]")
-                for img in images:
-                    lastMeasure.before_measure_imgs.create(img=img)
+            if f"{inspectionResult.pk}-content" in request.POST.keys():
+                if lastMeasure.isCM:
+                    newMeasure = inspectionResult.before_measure.create(
+                        content=request.POST[f"{inspectionResult.pk}-content"],
+                    )
+                    images = request.FILES.getlist(f"{inspectionResult.pk}-images[]")
+                    for img in images:
+                        newMeasure.before_measure_imgs.create(img=img)
+                else:
+                    lastMeasure.content = request.POST[f"{inspectionResult.pk}-content"]
+                    lastMeasure.save()
+                    before_images = list(
+                        request.POST.getlist(
+                            f"{inspectionResult.pk}-images-preloaded[]"
+                        )
+                    )
+                    for measure_img in lastMeasure.before_measure_imgs.all():
+                        if str(measure_img.pk) not in before_images:
+                            measure_img.delete()
+                    images = request.FILES.getlist(f"{inspectionResult.pk}-images[]")
+                    for img in images:
+                        lastMeasure.before_measure_imgs.create(img=img)
         messages.success(request, "저장이 완료되었습니다.")
         isSave = True
     return render(
@@ -323,10 +305,14 @@ def measure_before_install_service(request, urlCode):
     )
 
 
-def success_before_install_checklist_service(pk):
+def success_before_install_checklist_service(request, pk):
     checklist = get_object_or_404(BeforeInstallCheckList, pk=pk)
     checklist.isSuccess = True
     checklist.isCheckCM = True
     checklist.isCheckWriter = True
+    link = request.build_absolute_uri(
+        f"/work/read_before_install/{checklist.equipment}/{pk}",
+    )
+    sms_send(link, [checklist.cm.phone], 4)
     checklist.save()
     return JsonResponse({"result": "success"})
